+++ src/setup.ts
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import { fs, types, selectors, util } from 'vortex-api';
 import { isJunctionDir, createJunction, removeJunction } from './util';
 import path from 'path';
@@ -5,6 +6,8 @@ import path from 'path';
 import { setDirectoryJunctionEnabled } from './actions/settings';
 import { GAME_ID } from './common';
 import { IJunctionProps } from './types';
+
+import { migrateExtension } from './migrations/migrations';
 
 // This code executes when the user first manages Starfield AND each time they swap from another game to Starfield. 
 export async function setup(api: types.IExtensionApi,
@@ -20,6 +23,7 @@ export async function setup(api: types.IExtensionApi,
   }
   // Make sure the folder exists
   await fs.ensureDirAsync(myGamesFolder);
+  await migrateExtension(api);
 }
 
 export async function toggleJunction(api: types.IExtensionApi, enable: boolean): Promise<void> {